requirements.txt
# ML Tree System Dependencies
# ===========================

# Core ML
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Model persistence
joblib>=1.3.0

# Configuration
pyyaml>=6.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.24.0  # For FastAPI testing

# Development (optional)
# ipython>=8.0.0
# jupyter>=1.0.0
